[1825] Contribute selective HTML generation.
This contribution adds a new contextual menu entry on Capella models to
generate the HTML documentation of only *some* elements.
Compared to the existing dialog, a new first page is added to present
the user with a tree view of the model, allowing them to select for
which elements they wish to generate the documentation. This selection
can also be saved/loaded to a file in order to ease re-generating the
documentation for a specific set of elements.

Bug: 1825
Change-Id: Ibbc543f62b14408a77fd432143f815df52ea45c7
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>