[1825] Contribute selective HTML generation.
This contribution adds a new contextual menu entry on Capella models to
generate the HTML documentation of only *some* elements.
Compared to the existing dialog, a new first page is added to present
the user with a tree view of the model, allowing them to select for
which elements they wish to generate the documentation. This selection
can also be saved/loaded to a file in order to ease re-generating the
documentation for a specific set of elements.

Bug: 1825
Change-Id: I66ab7addabe99853c39ed0112bae1ef2906c2812
Signed-off-by: Florent Latombe <florent.latombe@obeo.fr>