[566977] URLS with Diagrams anchors redirection
Once the HTML has been generated, it is possible to see a list of all
diagrams for a given engineering phase (OA/SA/LA/PA) by clicking on each
engineering level folder.
From this list, it is possible to navigate to each diagram.
But, if instead of navigating, you copy this URL and paste it into a new
tab in your browser, then it would re-create the frame but not point
exactly to the anchor where the diagram is in the page, but rather
display the top of the page.
Basically, the reason for this is because there is a javascript
instruction on each main page that redirect to the top frame if someone
tries to display the page without the frames. But the anchor is lost in
the process.

Bug: 566977

Change-Id: I2781ae08f66108599edbafe366894c7c59680c7f
Signed-off-by: Axel RICHARD <axel.richard@obeo.fr>