#211 Remove non relevant content from test result html content
- Remove coord attribute in area tag
The content of the coord attribute generated for cliquable areas of
diagrams is sensitive depending on the docgen execution environment.
This attribute content is to be tested in a dedicated api test.
Update expected test results accordingly.

- Remove all attributes content in area tag
The content of the map tag generated for clickable areas of
diagrams is sensitive depending on the docgen execution environment.
Size of elements variations depending on execution environment may lead
to different order of the area elements.
Update expected test results accordingly.

Change-Id: I28128e7dc8e67b5b4ef2ff19d3675ccc9136b6cc
Signed-off-by: Arnaud Dieumegard <arnaud.dieumegard@obeo.fr>