#201 Check Capella version
Add new CheckCapellaVersion task launched at the beggining of the
documentaiton generation. If task founds that Capella version does not
match the docgen version then it cancels the generation.

Updated management of the core and configuration CommanLine implementations
to handle the generation failing through CheckCapellaVersion task. Log
is diplayed in the commandline and fails the complete generation.

Change-Id: I0000000000000000000000000000000000000000
Signed-off-by: Arnaud Dieumegard <arnaud.dieumegard@obeo.fr>